+++ src/main/java/com/dic/bill/model/scott/Anabor.java
@@ -1,17 +1,10 @@
 package com.dic.bill.model.scott;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-
 import com.dic.bill.Compress;
-
 import lombok.Getter;
 import lombok.Setter;
+
+import javax.persistence.*;
 
 /**
  * Архив наборов услуг по периодам
@@ -30,7 +23,7 @@ public class Anabor implements java.io.Serializable, Compress {
 
     @Id
 	@Column(name = "id", updatable = false, nullable = false)
-	private Integer id; // Id
+	private Long id; // Id
 
 	// лиц.счет
 	@ManyToOne(fetch = FetchType.LAZY)
